Switch to 32 bit hashes

+++ src/hashtable/hashtable.h
@@ -22,7 +22,7 @@ extern "C" {
 
 typedef uint8_t hashtable_bucket_key_value_flags_t;
 typedef uint32_t hashtable_bucket_hash_t;
-typedef uint64_t hashtable_bucket_index_t;
+typedef uint32_t hashtable_bucket_index_t;
 typedef hashtable_bucket_index_t hashtable_bucket_count_t;
 typedef uint32_t hashtable_key_size_t;
 typedef char hashtable_key_data_t;
